+++ page-templates/homepage.php
@@ -9,5 +9,6 @@
 get_header();
 
 get_template_part('template-parts/components/home-hero-slider');
+get_template_part('template-parts/components/interactive-select');
 
 get_footer();